Refactor older puzzles to be triggered by tests
There's a bug with Kotlin/IntelliJ that I encountered on stream, which causes
the code analysis and compilation within IntelliJ to not run if there's
more than one `main()` function in the project. Why this didn't trigger
until literally the TWELFTH main method is a mystery. But the way to fix this
is to trigger the puzzles the way God intended -- from the tests. If you want
more details on the bug, you can find the YouTrack issue at:
https://youtrack.jetbrains.com/issue/IDEA-258093.

For upcoming puzzles (as well as going back and changing Day 11 in a future
commit), we'll be triggering it from the tests which will be better refactored.
For legacy problems (1 through 10), we simply call them all from a single
test class which bluntly goes through each problem and calls the top-level
method.